third_party/WebKit/LayoutTests/fast/dom/Window/window-legacy-event-listener.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<script>
description("Test addEventListener() and removeEventListener() fail silently if arguments are missing.");

shouldThrow('window.addEventListener("foo")');
shouldThrow('window.removeEventListener("bar")');

shouldThrow('window.addEventListener()');
shouldThrow('window.removeEventListener()');
</script>
</body>
</html>
